# GH-3572 Concurrent WeakObjectRegistry