--- a/core/sail/memory/src/main/java/org/eclipse/rdf4j/sail/memory/MemorySailStore.java
+++ b/core/sail/memory/src/main/java/org/eclipse/rdf4j/sail/memory/MemorySailStore.java
@@ -40,7 +40,6 @@ import org.eclipse.rdf4j.sail.base.SailDataset;
 import org.eclipse.rdf4j.sail.base.SailSink;
 import org.eclipse.rdf4j.sail.base.SailSource;
 import org.eclipse.rdf4j.sail.base.SailStore;
-import org.eclipse.rdf4j.sail.memory.model.CloseableIterator;
 import org.eclipse.rdf4j.sail.memory.model.MemBNode;
 import org.eclipse.rdf4j.sail.memory.model.MemIRI;
 import org.eclipse.rdf4j.sail.memory.model.MemResource;
@@ -51,6 +50,7 @@ import org.eclipse.rdf4j.sail.memory.model.MemStatementList;
 import org.eclipse.rdf4j.sail.memory.model.MemTriple;
 import org.eclipse.rdf4j.sail.memory.model.MemValue;
 import org.eclipse.rdf4j.sail.memory.model.MemValueFactory;
+import org.eclipse.rdf4j.sail.memory.model.WeakObjectRegistry;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -785,7 +785,8 @@ class MemorySailStore implements SailStore {
 			Lock stLock = openStatementsReadLock();
 			try {
 				int snapshot = getCurrentSnapshot();
-				try (CloseableIterator<MemIRI> memIRIsIterator = valueFactory.getMemIRIsIterator()) {
+				try (WeakObjectRegistry.AutoCloseableIterator<MemIRI> memIRIsIterator = valueFactory
+						.getMemIRIsIterator()) {
 					while (memIRIsIterator.hasNext()) {
 						MemResource memResource = memIRIsIterator.next();
 						if (isContextResource(memResource, snapshot)) {
@@ -794,7 +795,8 @@ class MemorySailStore implements SailStore {
 					}
 				}
 
-				try (CloseableIterator<MemBNode> memBNodesIterator = valueFactory.getMemBNodesIterator()) {
+				try (WeakObjectRegistry.AutoCloseableIterator<MemBNode> memBNodesIterator = valueFactory
+						.getMemBNodesIterator()) {
 					while (memBNodesIterator.hasNext()) {
 						MemResource memResource = memBNodesIterator.next();
 						if (isContextResource(memResource, snapshot)) {