--- a/core/sail/memory/src/main/java/org/eclipse/rdf4j/sail/memory/model/CloseableIterator.java
+++ b/core/sail/memory/src/main/java/org/eclipse/rdf4j/sail/memory/model/CloseableIterator.java
@@ -1,0 +1,17 @@
+/*******************************************************************************
+ * Copyright (c) 2022 Eclipse RDF4J contributors.
+ *  All rights reserved. This program and the accompanying materials
+ *  are made available under the terms of the Eclipse Distribution License v1.0
+ *  which accompanies this distribution, and is available at
+ *  http://www.eclipse.org/org/documents/edl-v10.php.
+ ******************************************************************************/
+
+package org.eclipse.rdf4j.sail.memory.model;
+
+import java.util.Iterator;
+
+public interface CloseableIterator<E> extends Iterator<E>, AutoCloseable {
+
+	@Override
+	void close();
+}