GitHub issue resolved: #3572 <!-- add a Github issue number here, e.g #123. -->

Briefly describe the changes proposed in this PR:
 - use an array of WeakHashMap objects to allow multithreaded access for both read and writes
     - we mod the hash of the key to index into the array
     - initial size of a WeakHashMap is 16 and we create `Runtime.getRuntime().availableProcessors() * 2` maps
         - this gives some overhead for MemoryStores with very very few values
         - as the maps fill up there should be less and less overhead since we are storing the same number of objects in total
 - use StampedLocks to allow multiple readers and single writer per WeakHashMap
     - this allows multiple writers if we hit different WeakHashMaps
 - remove external synchronization
 - use lambdas to support "getOrAdd" operations against the WeakObjectRegistry
 - deprecate three methods that are not thread-safe and create new methods that are thread-safe to achieve the same goal

<!-- short description of your change goes here -->

----
PR Author Checklist (see the [contributor guidelines](https://github.com/eclipse/rdf4j/blob/main/.github/CONTRIBUTING.md) for more details):

 - [ ] my pull request is [self-contained](https://rdf4j.org/documentation/developer/merge-strategy/#self-contained-changes-pull-requests-and-commits)
 - [ ] I've added tests for the changes I made
 - [ ] I've applied [code formatting](https://github.com/eclipse/rdf4j/blob/main/.github/CONTRIBUTING.md#code-formatting) (you can use `mvn process-resources` to format from the command line)
 - [ ] I've [squashed](https://rdf4j.org/documentation/developer/squashing) my commits where necessary 
 - [ ] every commit message starts with the issue number (GH-xxxx) followed by a meaningful description of the change