@hmottestad 2022-01-05
### Before
```
Benchmark                                      Mode  Cnt    Score     Error  Units
MemValueFactoryConcurrentBenchmark.onlyReads   avgt    5  491.734 ± 110.388  ms/op
MemValueFactoryConcurrentBenchmark.onlyWrites  avgt    5  707.663 ±  79.292  ms/op
MemValueFactoryConcurrentBenchmark.readHeavy   avgt    5  217.618 ±   7.088  ms/op
```

### After
```
```