@hmottestad 2022-01-05
### Before
```
Benchmark                                                     Mode  Cnt      Score   Error  Units
QueryBenchmark.complexQuery                                   avgt          49.555          ms/op
QueryBenchmark.different_datasets_with_similar_distributions  avgt           3.250          ms/op
QueryBenchmark.groupByQuery                                   avgt           2.082          ms/op
QueryBenchmark.long_chain                                     avgt        1023.599          ms/op
QueryBenchmark.lots_of_optional                               avgt         259.100          ms/op
QueryBenchmark.minus                                          avgt       13122.140          ms/op
QueryBenchmark.nested_optionals                               avgt         259.591          ms/op
QueryBenchmark.pathExpressionQuery1                           avgt          22.957          ms/op
QueryBenchmark.pathExpressionQuery2                           avgt           1.815          ms/op
QueryBenchmark.query_distinct_predicates                      avgt         172.088          ms/op
QueryBenchmark.simple_filter_not                              avgt          10.976          ms/op

```

### After

```
```