@hmottestad 2022-01-06
### Before
```
```

### After
```
Benchmark                    (isolationLevel)  Mode  Cnt    Score    Error  Units
NotifyingSailBenchmark.load              NONE  avgt    5  157.878 ± 22.213  ms/op
NotifyingSailBenchmark.load          SNAPSHOT  avgt    5  198.236 ± 22.420  ms/op
NotifyingSailBenchmark.load      SERIALIZABLE  avgt    5  392.916 ± 25.133  ms/op
```