@hmottestad 2022-01-06
So this is adding a small overhead when loading data into a MemoryStore. I'll investigate.