@hmottestad 2022-01-06
### Before
```
Benchmark                                      Mode  Cnt    Score    Error  Units
MemValueFactoryConcurrentBenchmark.onlyReads   avgt    5  450.121 ± 78.835  ms/op
MemValueFactoryConcurrentBenchmark.onlyWrites  avgt    5  749.964 ± 62.857  ms/op
MemValueFactoryConcurrentBenchmark.readHeavy   avgt    5  217.716 ± 11.002  ms/op
```

### After
```
Benchmark                                      Mode  Cnt    Score    Error  Units
MemValueFactoryConcurrentBenchmark.onlyReads   avgt    5   75.562 ±  1.790  ms/op
MemValueFactoryConcurrentBenchmark.onlyWrites  avgt    5  416.135 ± 16.399  ms/op
MemValueFactoryConcurrentBenchmark.readHeavy   avgt    5   83.085 ±  0.777  ms/op
```