@hmottestad 2022-01-06
The last benchmarks are specifically designed to show the performance improvements with this PR. That benchmark has three benchmarks that simulate various access patterns against the MemValyeFactory using multiple threads. The other benchmarks I've posted are just to show that there is no significant performance degradation for queries or when adding data.

Due to other bottlenecks in the system there is little performance improvements elsewhere due to this patch. One of those bottlenecks is locking, which I'm trying to address in the following PR https://github.com/eclipse/rdf4j/pull/3564