@hmottestad 2022-01-08
```
Benchmark                                      Mode  Cnt    Score   Error  Units
MemValueFactoryConcurrentBenchmark.onlyReads   avgt   10   77.343 ± 0.272  ms/op
MemValueFactoryConcurrentBenchmark.onlyWrites  avgt   10  437.942 ± 3.844  ms/op
MemValueFactoryConcurrentBenchmark.readHeavy   avgt   10   84.729 ± 0.071  ms/op
```